Change if_output interface to allow passing cached lle   down to L2: 186574, 186577, 186673
186574:
- Change if_output interface to allow passing cached lle
  down to L2

186577:
- used cache lle where possible

186673:
- fix if_output calls in modules

+++ sys/contrib/ipfilter/netinet/ip_compat.h
@@ -1825,7 +1825,7 @@ typedef	struct	tcpiphdr	tcpiphdr_t;
 #endif
 
 #ifndef offsetof
-# define offsetof(t,m) (int)((&((t *)0L)->m))
+# define offsetof(t,m) (size_t)((&((t *)0L)->m))
 #endif
 
 /*
